requirements.txt
djangorestframework==3.15.2
ipython==8.25.0